pyproject.toml
[project]
name = "pis-optimizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "attrs>=24.2.0",
    "jmetalpy>=1.7.0",
    "pandas>=2.2.3",
    "ruff>=0.8.3",
]
